- moved winesList component to its own folder - added a Route for a single view of a wine
