--- client/components/SingleWine/index.js
@@ -1,1 +1,0 @@
-export default require('./SingleWineContainer')
