+++ client/components/WineList/winesIcon.spec.js
