+++ client/components/WineList/winesList.spec.js
@@ -6,7 +6,7 @@ import enzyme, {shallow} from 'enzyme'
 import Adapter from 'enzyme-adapter-react-16'
 import {WinesList} from './winesList'
 import WinesIcon from './winesIcon'
-import store from '../store'
+import store from '../../store'
 
 const adapter = new Adapter()
 enzyme.configure({adapter})
