+++ client/components/index.js
@@ -6,4 +6,5 @@
 export {default as Navbar} from './navbar'
 export {default as UserHome} from './user-home'
 export {Login, Signup} from './auth-form'
-export {default as WinesList} from './winesList'
+export {default as SingleWine} from './SingleWine/SingleWineContainer'
+export {default as WinesList} from './WineList/winesList'
