+++ client/routes.js
@@ -2,7 +2,7 @@ import React, {Component} from 'react'
 import {connect} from 'react-redux'
 import {withRouter, Route, Switch} from 'react-router-dom'
 import PropTypes from 'prop-types'
-import {Login, Signup, UserHome, WinesList} from './components'
+import {Login, Signup, UserHome, WinesList, SingleWine} from './components'
 import {me} from './store'
 
 /**
@@ -19,6 +19,7 @@ class Routes extends Component {
       <Switch>
         {/* Routes placed here are available to all visitors */}
         <Route exact path="/" component={WinesList} />
+        <Route exact path="/wines/:wineId" component={SingleWine} />
         <Route path="/login" component={Login} />
         <Route path="/signup" component={Signup} />
         {isLoggedIn && (
